config: add pmemkv_config_put_object_cb function
It allows passing object with additional state to config. Pointer which is passed
to pmemkv_config_put_object_cb does not have to be of a specific type (like in
pmemkv_config_put_object). Instead user should provide a callback getter function which
interprets the custom object and returns a pointer which is expected by pmemkv.

It is used in C++ bindings to implement `put(std::string key, std::unique_ptr value)`.